Day 6 and switch to i64 for larger results

+++ src/day_1/part1.rs
@@ -1,4 +1,4 @@
-pub fn part1(lines: Vec<String>) -> i32 {
+pub fn part1(lines: Vec<String>) -> i64 {
   let readings = lines.iter().map(|line| line.parse::<i32>().unwrap()).collect::<Vec<i32>>();
   let pairs = readings.iter().zip(readings.iter().skip(1));
   pairs.filter(|(reading1, reading2)| reading1 < reading2).count().try_into().unwrap()
