+++ src/day_2/part1.rs
@@ -5,7 +5,7 @@ struct Position {
   depth: i32,
 }
 
-pub fn part1(lines: Vec<String>) -> i32 {
+pub fn part1(lines: Vec<String>) -> i64 {
   let instructions = lines.iter().map(|line| {
     let parts = line.split(" ").collect::<Vec<&str>>();
     let instruction = match parts[0] {
@@ -24,5 +24,5 @@ pub fn part1(lines: Vec<String>) -> i32 {
       Instruction::Up => Position { horizontal: position.horizontal, depth: position.depth - command.value },
     }
   });
-  final_position.horizontal * final_position.depth
+  (final_position.horizontal * final_position.depth) as i64
 }
